// admin/src/api/generated/schemas/adminAccountingAccountTypeItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminAccountingAccountTypeItem = {
  code?: string;
  name?: string;
  normal_side?: string;
  asset_codes?: string[];
  description?: string;
};